rename package properly
