+++ src/main/java/de/unistuttgart/iste/meitrex/assignment_service/config/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package should contain any classes that are used to configure the application.
  */
-package de.unistuttgart.iste.meitrex.template.config;
+package de.unistuttgart.iste.meitrex.assignment_service.config;
