+++ src/main/java/de/unistuttgart/iste/meitrex/assignment_service/controller/TemplateController.java
@@ -1,7 +1,7 @@
-package de.unistuttgart.iste.meitrex.template.controller;
+package de.unistuttgart.iste.meitrex.assignment_service.controller;
 
 import de.unistuttgart.iste.meitrex.generated.dto.Template;
-import de.unistuttgart.iste.meitrex.template.service.TemplateService;
+import de.unistuttgart.iste.meitrex.assignment_service.service.TemplateService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
