+++ src/main/java/de/unistuttgart/iste/meitrex/assignment_service/persistence/entity/TemplateEntity.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.iste.meitrex.template.persistence.entity;
+package de.unistuttgart.iste.meitrex.assignment_service.persistence.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
