+++ src/main/java/de/unistuttgart/iste/meitrex/assignment_service/persistence/mapper/package-info.java
@@ -1,4 +1,4 @@
 /**
  * This package contains all the mapper classes that map between DTOs and entities.
  */
-package de.unistuttgart.iste.meitrex.template.persistence.mapper;
+package de.unistuttgart.iste.meitrex.assignment_service.persistence.mapper;
