+++ src/main/java/de/unistuttgart/iste/meitrex/assignment_service/validation/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Put class level validation logic in this package.
  */
-package de.unistuttgart.iste.meitrex.template.validation;
+package de.unistuttgart.iste.meitrex.assignment_service.validation;
