+++ src/test/java/de/unistuttgart/iste/meitrex/assignment_service/TemplateForMicroservicesApplicationTests.java
@@ -1,4 +1,4 @@
-package de.unistuttgart.iste.meitrex.template;
+package de.unistuttgart.iste.meitrex.assignment_service;
 
 import org.springframework.boot.test.context.SpringBootTest;
